Almost fin

+++ src/include/heap.h
@@ -2,24 +2,24 @@
 #define HEAP_H_
 
 typedef struct {
-    uint8_t key; //для приоритета в очереди
-    uint8_t value; //значение ребра
+    int key; //для приоритета в очереди
+    int value; //значение ребра
 } HeapNode;
 
 typedef struct {
-    uint8_t max_size;
-    uint8_t num_nodes;
+    int max_size;
+    int num_nodes;
 
     HeapNode *nodes;
 } Heap;
 
 void free_heap(Heap* buf); //
 HeapNode *heap_min(Heap* buf); //
-Heap *heap_create(uint8_t max_size); //
+Heap *heap_create(int max_size); //
 void heap_extract_min(Heap* buf); //
-void heapify(Heap* buf, uint8_t index); //
+void heapify(Heap* buf, int index); //
 void heap_swap(HeapNode* a, HeapNode* b); //
-void heap_insert(Heap* buf, uint8_t key, uint8_t value); //
-int heap_decrease_key(Heap* buf, uint8_t value, uint8_t new_key, uint8_t size); //
+void heap_insert(Heap* buf, int key, int value); //
+int heap_decrease_key(Heap* buf, int value, int new_key, int size); //
 
 #endif //HEAP_H_
